+++ src/Payment/CC.php
@@ -2,6 +2,7 @@
 
 namespace Wearesho\Bobra\Platon\Payment;
 
+use Wearesho\Bobra\Payments\PayerDetailsInterface;
 use Wearesho\Bobra\Payments\UrlPairInterface;
 use Wearesho\Bobra\Platon;
 
@@ -16,6 +17,7 @@ class CC extends Platon\Payment
         int $id,
         string $lang,
         UrlPairInterface $urlPair,
+        PayerDetailsInterface $payerDetails,
         string $sign,
         string $key,
         string $formUrl,
@@ -24,7 +26,7 @@ class CC extends Platon\Payment
         string $formId = null,
         string $cardToken = null
     ) {
-        parent::__construct($id, $lang, $urlPair, $sign, $key, $formUrl, $ext, $formId, $cardToken);
+        parent::__construct($id, $lang, $urlPair, $payerDetails, $sign, $key, $formUrl, $ext, $formId, $cardToken);
         $this->data = $data;
     }